pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "consensus-dpo"
version = "0.1.0"
description = "Consensus → DPO pipeline with Novita AI GPT-OSS"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{ name = "SrivastavaNik", email = "" }]
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "redis>=5.0.4",
  "tenacity>=8.4.1",
  "aiolimiter>=1.1.0",
  "mlflow>=2.13.0",
  "numpy>=1.26.0",
  "orjson>=3.10.1"
]

[project.optional-dependencies]
train = [
  "torch>=2.3.0; platform_system != 'Darwin' or platform_machine != 'arm64'",
  "transformers>=4.42.0",
  "trl>=0.8.6",
  "accelerate>=0.31.0",
  "peft>=0.11.1",
  "bitsandbytes>=0.43.1; platform_system != 'Darwin'",
  "datasets>=2.19.0"
]
retrieval = [
  "faiss-cpu>=1.8.0"
]
eval = [
  "evaluate>=0.4.2",
  "scikit-learn>=1.5.0"
]

[tool.hatch.build.targets.wheel]
packages = [
  "libs/consensus_dpo"
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B"]


